ui/utils: Don't allude to gnome-desktop's clock format
We're no longer trying to match gnome-desktop timestamps now, so drop
the comment related to this in the `ChronoExt` code. Also rename the
function from `to_gnome_format()` back to `to_locale()` as it was before
(`%x` is locale-sensitive, and while the 12h/24h setting isn't strictly
part of the locale, it's still strongly related).

Follow-up to 911b735 ("archives_page: respect 12h/24h setting when
displaying timestamps").