Avoid allocating in `pre_exec` closure
`std::io::Error::other` allocates memory (see [rust-lang/rust#148971]).
This is bad in multi-threaded programs, which pika-backup AFAIK is. If
the fork occurs while the allocator lock is held by another thread,
deadlocks can occur, since there's no one left in the new process to
unlock the mutex. I do not believe this is UB, and modern libc offer
protections against this issue, but this isn't POSIX-compliant and
should preferably be avoided.

`nix` provides a non-allocating `impl From<Errno> for std::io::Error`,
which can be used instead. Even if `std::io::Error::other` was
non-allocating, using this impl is better because it preserves OS error
code instead of forwarding a placeholder EINVAL to parent process.

[rust-lang/rust#148971]: https://github.com/rust-lang/rust/pull/148971